+++ projects/elonkit/src/ui/action-heading/__stories__/action-heading-story-click/action-heading-story-click.component.ts
@@ -7,7 +7,7 @@ import { Component, ChangeDetectionStrategy } from '@angular/core';
   changeDetection: ChangeDetectionStrategy.OnPush
 })
 export class ActionHeadingStoryClickComponent {
-  text = 'Hello World';
+  public text = 'Hello World';
 
   public alert() {
     alert('Its working!');