+++ projects/elonkit/src/ui/alert/alert.component.locale.ts
@@ -2,10 +2,10 @@ import { Injectable } from '@angular/core';
 
 @Injectable({ providedIn: 'root' })
 export class ESAlertLocale {
-  labelClose = 'Close';
+  public labelClose = 'Close';
 }
 
 @Injectable()
 export class ESAlertLocaleRU extends ESAlertLocale {
-  labelClose = 'Закрыть';
+  public labelClose = 'Закрыть';
 }