+++ projects/elonkit/src/ui/breadcrumbs/__specs__/breadcrumbs.spec.service.ts
@@ -67,22 +67,22 @@ const ITEMS = [
 
 @Injectable()
 export class CategoriesService {
-  getAll() {
+  public getAll() {
     return of(CATEGORIES);
   }
 
-  getOne(id: number) {
-    return of(CATEGORIES.find(item => item.id === id));
+  public getOne(id: number) {
+    return of(CATEGORIES.find((item) => item.id === id));
   }
 }
 
 @Injectable()
 export class ItemsService {
-  getAll(categoryId: number) {
-    return of(ITEMS.filter(item => item.categoryId === categoryId));
+  public getAll(categoryId: number) {
+    return of(ITEMS.filter((item) => item.categoryId === categoryId));
   }
 
-  getOne(id: number) {
-    return of(ITEMS.find(item => item.id === id));
+  public getOne(id: number) {
+    return of(ITEMS.find((item) => item.id === id));
   }
 }