+++ projects/elonkit/src/ui/dropzone/__stories__/dropzone-story-custom-icon/dropzone-story-custom-icon.component.ts
@@ -1,0 +1,23 @@
+import { Component, ChangeDetectionStrategy } from '@angular/core';
+import { FormGroup, FormControl } from '@angular/forms';
+import { DomSanitizer } from '@angular/platform-browser';
+import { MatIconRegistry } from '@angular/material/icon';
+
+@Component({
+  selector: 'es-dropzone-custom-icon',
+  templateUrl: './dropzone-story-custom-icon.component.html',
+  styleUrls: ['./dropzone-story-custom-icon.component.scss'],
+  changeDetection: ChangeDetectionStrategy.OnPush
+})
+export class DropzoneStoryCustomIconComponent {
+  public form = new FormGroup({
+    docs: new FormControl([])
+  });
+
+  constructor(private iconRegistry: MatIconRegistry, private sanitizer: DomSanitizer) {
+    iconRegistry.addSvgIcon(
+      'upload',
+      sanitizer.bypassSecurityTrustResourceUrl('/icons/upload.svg')
+    );
+  }
+}