+++ projects/elonkit/src/ui/dropzone/__stories__/dropzone-story-required/index.ts
@@ -1,0 +1,3 @@
+export { DropzoneStoryRequiredModule } from './dropzone-story-required.module';
+export { DropzoneStoryRequiredComponent } from './dropzone-story-required.component';
+export { DROPZONE_STORY_REQUIRED_SOURCE } from './dropzone-story-required.source';