+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - [ActionHeading] Introduce new component
 - [Alert] Introduce new component
 - [Breadcrumbs] Introduce new component
+- [Dropzone] Introduce new component
 - [EmptyState] Introduce new component
 - [InlineFormField] Introduce new component
 - [Paginator] Introduce new component